Improve error handling UX for insertBefore and removeChild client exception errors
