+++ apps/studio/components/ui/ErrorBoundary/ErrorBoundary.tsx
